Created routers
The functions are just fill functions. They're not real. You can create them as fit, they're just to show what goes in as an example OR you can just do it as it's done in our server code. I just established the endpoints for EVERYTHING that was in our RESTful model. If you're confused check out the store-server code.